Extract the tests of common EntryIndex functionality in the base class
